feat: Add support for Todoist and Google Tasks; update configuration options
