+++ Config.ts
@@ -26,6 +26,9 @@ export interface MutableConfig {
     max_threads: number;
     default_task_list_name: string;
     GEMINI_API_KEY: string;
+    task_service: string;
+    todoist_api_key: string;
+    todoist_project_id: string;
 }
 
 export class Config implements Readonly<MutableConfig> {
@@ -38,6 +41,9 @@ export class Config implements Readonly<MutableConfig> {
     public readonly max_threads: number;
     public readonly default_task_list_name: string;
     public readonly GEMINI_API_KEY: string;
+    public readonly task_service: string;
+    public readonly todoist_api_key: string;
+    public readonly todoist_project_id: string;
 
     private static validate(config: Config) {
         Utils.assert(config.unprocessed_label.length > 0, "unprocessed_label can't be empty");
@@ -56,6 +62,9 @@ export class Config implements Readonly<MutableConfig> {
             max_threads: 50,
             default_task_list_name: "My Tasks",
             GEMINI_API_KEY: "",
+            task_service: "Google Tasks",
+            todoist_api_key: "",
+            todoist_project_id: "",
         };
 
         const values = Utils.withTimer("GetConfigValues", () => {
@@ -100,7 +109,10 @@ export class Config implements Readonly<MutableConfig> {
                 case "processing_failed_label":
                 case "go_link":
                 case "default_task_list_name":
-                case "GEMINI_API_KEY": {
+                case "GEMINI_API_KEY":
+                case "task_service":
+                case "todoist_api_key":
+                case "todoist_project_id": {
                     config[name] = value;
                     break;
                 }
