+++ Next.md
@@ -1,3 +1,1 @@
 Next tasks:
-
-  - New settings should be added to the configs sheet to let you select your preferred task service (`task_service`) and provide your Todoist API key (`todoist_api_key`) and project ID (`todoist_project_id`).
